+++ tests/InfluxDB/dashboard_static_custom_title.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+make DATASOURCE=influxdb \
+     POLICY=autogen \
+     MEASUREMENT=tarantool_http \
+     TITLE='My custom title' \
+     build-static-influxdb